add npm script for explorer in example project
