+++ apps/example/package.json
@@ -6,7 +6,8 @@
     "build": "tsc -p . --noEmit && tsup src/entrypoints/*.ts --format cjs",
     "start:http": "tsup src/entrypoints/http.ts --format cjs --onSuccess 'node dist/http.cjs'",
     "start:reactions": "tsup src/entrypoints/reactions.ts --format cjs --onSuccess 'node dist/reactions.cjs'",
-    "test": "vitest run"
+    "test": "vitest run",
+    "explorer": "NODE_PATH=./node_modules eqx-explorer -cs $MDB_RO_CONN_STR -c ./src/explorer-config.ts"
   },
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.410.0",
