--- apps/example/src/entrypoints/explorer-config.ts
@@ -1,11 +1,0 @@
-import { Invoice, InvoiceAutoEmailer, Payer } from "../domain/index.js"
-
-export const categories = [
-  { name: Invoice.Stream.CATEGORY, codec: Invoice.Events.codec, fold: Invoice.Fold },
-  { name: Payer.Stream.category, codec: Payer.Events.codec, fold: Payer.Fold },
-  {
-    name: InvoiceAutoEmailer.Stream.category,
-    codec: InvoiceAutoEmailer.Events.codec,
-    fold: InvoiceAutoEmailer.Fold,
-  },
-]
